Features/SS add cart logic

* added simple fileds

* Added api Get
    Added new Models

* Added promocode api

* Added basket/packaging api

* remove a few lines

* Fixed signature methods
    Remove unnecessary file

* wip

* Fixed returned signature

* Fixed Packaging and Cart dto/models

* fixed build

* wip

* wip

* rename class to ImageInfo
    Added item for cart product

* Rename class to ImageInfo

* wip

* Added ctor

* Remove product UId from CartDto
    Fixed models and Id long type

* Rename class CartTopping
    change type Id field
    Refactor code
    Fixed Mapper

* Rename Promocode
    Refactor code

* Remove methods Prepare

* Added new VM's for cart items

* Added new api /sauces /clear

* big changes)))

* fix loading cart

* wip

Approved-by: Alex Semelit <semelitto@gmail.com>

+++ SushiShop.Core/Data/Models/Cities/City.cs
@@ -6,7 +6,7 @@ namespace SushiShop.Core.Data.Models.Cities
     public class City
     {
         public City(
-            int id,
+            long id,
             string name,
             int shopsCount,
             double latitude,
@@ -29,7 +29,7 @@ namespace SushiShop.Core.Data.Models.Cities
             Currency = currency;
         }
 
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public int ShopsCount { get; set; }
         public double Latitude { get; set; }
